feat: add admin router and a lot of other fixings

+++ prova/src/main/java/br/ufrn/tads/prova/config/DataLoader.java
@@ -12,6 +12,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Configuration
@@ -82,7 +83,7 @@ public class DataLoader {
             // Produto 1
             Yatch y1 = new Yatch();
             y1.setName("Veleiro Ocean");
-            y1.setPrice(250000.0);
+            y1.setPrice(BigDecimal.valueOf(250000.0));
             y1.setColor("white");
             y1.setStatus(true);
             y1.setCodProduct("PROD-2222");
@@ -91,7 +92,7 @@ public class DataLoader {
             // Produto 2
             Yatch y2 = new Yatch();
             y2.setName("Lancha Speed");
-            y2.setPrice(150000.0);
+            y2.setPrice(BigDecimal.valueOf(150000.0));
             y2.setColor("white");
             y2.setStatus(true);
             y2.setCodProduct("PROD-3345");
@@ -100,7 +101,7 @@ public class DataLoader {
 // Produto 3
             Yatch y3 = new Yatch();
             y3.setName("Barco a Vela Breeze");
-            y3.setPrice(89000.0);
+            y3.setPrice(BigDecimal.valueOf(89000.0));
             y3.setColor("white");
             y3.setStatus(false);
             y3.setCodProduct("PROD-4489");
@@ -109,7 +110,7 @@ public class DataLoader {
 // Produto 4
             Yatch y4 = new Yatch();
             y4.setName("Iate Luxo");
-            y4.setPrice(750000.0);
+            y4.setPrice(BigDecimal.valueOf(750000.0));
             y4.setColor("white");
             y4.setStatus(true);
             y4.setCodProduct("PROD-5512");
@@ -118,7 +119,7 @@ public class DataLoader {
 // Produto 5
             Yatch y5 = new Yatch();
             y5.setName("Jet Ski Aqua");
-            y5.setPrice(32000.0);
+            y5.setPrice(BigDecimal.valueOf(32000.0));
             y5.setColor("white");
             y5.setStatus(true);
             y5.setCodProduct("PROD-6678");
@@ -127,7 +128,7 @@ public class DataLoader {
 // Produto 6
             Yatch y6 = new Yatch();
             y6.setName("Canoa Rápida");
-            y6.setPrice(12500.0);
+            y6.setPrice(BigDecimal.valueOf(12500.0));
             y6.setColor("white");
             y6.setStatus(false);
             y6.setCodProduct("PROD-7734");
@@ -136,7 +137,7 @@ public class DataLoader {
 // Produto 7
             Yatch y7 = new Yatch();
             y7.setName("Barco Pescador");
-            y7.setPrice(45000.0);
+            y7.setPrice(BigDecimal.valueOf(45000.0));
             y7.setColor("white");
             y7.setStatus(true);
             y7.setCodProduct("PROD-8891");
@@ -145,7 +146,7 @@ public class DataLoader {
 // Produto 8
             Yatch y8 = new Yatch();
             y8.setName("Inflável Explorer");
-            y8.setPrice(8900.0);
+            y8.setPrice(BigDecimal.valueOf(8900.0));
             y8.setColor("white");
             y8.setStatus(true);
             y8.setCodProduct("PROD-9903");
@@ -154,7 +155,7 @@ public class DataLoader {
 // Produto 9
             Yatch y9 = new Yatch();
             y9.setName("Catamarã Duo");
-            y9.setPrice(320000.0);
+            y9.setPrice(BigDecimal.valueOf(320000.0));
             y9.setColor("white");
             y9.setStatus(false);
             y9.setCodProduct("PROD-1015");
@@ -163,7 +164,7 @@ public class DataLoader {
 // Produto 10
             Yatch y10 = new Yatch();
             y10.setName("Hovercraft Turbo");
-            y10.setPrice(420000.0);
+            y10.setPrice(BigDecimal.valueOf(420000.0));
             y10.setColor("white");
             y10.setStatus(true);
             y10.setCodProduct("PROD-1127");
